Replace R_XXX values with REB_R "pseudotypes"
In an earlier commit, the ability was added to return an arbitrary
REBVAL* from a native's dispatcher...including API values which would
be released.  But it put special return results in-band by means of
the first byte, to get the special instructions to the evaluator
(as well as fast convenience routines that could produce nulls, logics,
blanks, etc. without having to run a full Move_Value).

Since then, it has become accepted to use "pseudotypes" for other
things--these are values above REB_MAX that are not contentious with
things like REB_BLOCK or REB_INTEGER.  This commit changes the return
instructions to use pseudotypes to signal the out of band states,
meaning what is returned is always a REBVAL*.

It also shifts EVAL to always use a subframe.  This gets rid of the
need for a pseudotype returned only by EVAL, and better reflects the
stack state that would exist if a user re-implemented the evaluator
by hand in a function called EVAL, by leaving the native on the stack.