Eliminate REB_R_REFERENCE payload in favor of frame union
Most "pseudotypes" used to return instructions to Eval_Core() and
Next_Path_Throws() did not have payloads; REB_R_REFERENCE being the
only exception.  One problem with that, however, is that if you use
the output cell to write the instruction's data then it will then
appear that the output pointer matches the frame's output...which
would otherwise indicate the "optimal case" of output being where it
needs to be.  Yet having a pseudotype in that spot means a switch() or
if() needs to be used.

One alternative is to write the payload for the pseudotype into the
frame's cell.  However, that could corrupt otherwise useful information
that is actually a real value and not a pseudotype.  For instance, the
path processing was using the frame's cell for the calculated picker
of the next step...which helped make sure things like evaluated GROUP!s
would be protected from GC.

The REB_R_REFERENCE pseudotype also was really just extracting the data
to put in frame fields, but lying because there weren't datatypes in
the frame that were compatible.  This bites the bullet and turns three
pointers worth of space in the frame into a union, so that Eval_Core()
and Next_Path_Throws() can accurately and clearly show what they are
doing with the data.  It's more direct and prevents having to worry
about where the pseudotypes write their information.