Simplify pseudotypes, reduce switch() overhead for common results
The conversion of the C `int` integer-based return values from
dispatchers to the idea of using REBVAL* involved the use of REB_XXX
values higher than the maximum.  But it also opened a new possibility
for a way of more optimally passing back values, via the special
recognition of a return of the frame's output cell...and the use
of functions that pipe their input to their output:

    return Init_Logic(D_OUT, x != y);

Since that not only initializes the cell in D_OUT but also returns it,
it seems easy enough to get rid of the idea of needing a switch()
statement on value types that aren't pseudotypes.  Additionally, the
construct:

    return Move_Value(D_OUT, some_value);

Can be simplified with a macro, and it seems there's not any particular
reason not to just call that RETURN in the internal core code:

    RETURN (some_value);

With this in place, it means it's not too bad to make a rule to say
that non-API handles should not be returned from natives unless they
are specifically the D_OUT pointer, and that thrown values are only
tolerated in D_OUT--since they're not supposed to be seen directly
by the API.  This makes for faster checking in most cases without the
need for a switch statement.