Eliminate native usage of frame's cell (via D_CELL)
Originally, the idea that each frame needed one cell's worth of
allocation came out of the need to be able to have a place to put a
generated value value during an EVAL operation.  Not only did it need
a location to be written into, but it also had to be GC protected:

    eval (func [] [recycle | 'print]) "What keeps func alive?"

Eventually, designing Eval_Core() to be re-entered allowed eval to
keep its argument alive by leaving it exactly where it was...in the
argument slot of the frame of the EVAL call.  Yet in the meantime,
the frame's cell had been exported as "D_CELL"...to be visible as a way
for native actions to get a scratch cell "for free" that was GC safe.

But unlike EVAL, a non-negotiable need for a non-movable per-frame
cell of storage space arose with the SHOVE operation.  The right hand
side of the operator can only be evaluated once, and its product must
be able to outlive evaluations on the left hand side.  Putting it on
the data stack is not an option, as it would be movable and the
f->gotten pointer could not stay fixed to it...not to mention needing
to adjust the data stack pointer to account for it during another
evaluation.

This removes D_CELL, and takes away the idea that natives are allowed
access to the frame's cell...reserving it for the evaluator's internal
usage.  Instead, it uses locals pushed with PUSH_GC_GUARD().

Includes improvements to catch unbalanced GC guards correctly, as the
feature of checking it had atrophied from lack of use.

Also removes the EVERY operation, as it was a client of the D_CELL that
had become broken.  It would involve thinking to fix it, and removing
it entirely seems better than leaving it broken.  See:

https://github.com/metaeducation/ren-c/issues/847