Resolve bugs in potentially "stale" lookback handling
The evaluator is supposed to have a narrow window of observation,
with frames advancing in a forward-only direction and examining only
one value at a time.  That value's address or stability can not be
taken for granted after the next value is fetched...so there was a
"lookback pointer" returned from the next fetch instruction.  The
system knows whether it is safe to use the same pointer, or to proxy
it into a safer location (e.g. FRM_CELL(f)).

There had been a feature for stress testing the lookback, forcibly
expiring pointers on each fetch.  This hadn't been used lately, and
revealed a bug introduced by a recent simplification of the SET-WORD!
and SET-PATH! code.

This commit changes each call to Fetch_Next_In_Frame() to have to
explicitly request whether it wants a lookback pointer or not, and
revisits the case where it might want one.  It also fixes up the debug
feature so that it no longer leaks on shutdown.