First take on SHOVE implementation (`->`)
This is a somewhat clumsy but workable first cut at a SHOVE operation.

The reason for the existence of SHOVE is to solve the historical
problem of not really having an answer for how to dispatch enfix
functions via path.  This causes problems when trying to use an enfix
operator out of a context, e.g.

    +: func [] [...] ;-- some local definition
    ...
    ...
    1 lib/+ 2 ;-- desire to use the definition in lib

This is probably the most compelling reason, but it is also true that
not being able to dispatch enfix from PATH! has meant that infix
functions in Rebol never were able to use refinements.

There are several reasons why "lookahead" through path evaluation is
not a very viable concept.  Lookahead on words is cheap by comparison,
and has no side effects...while PATH!s can contain GROUP!s.  A failed
lookahead would have to be able to disregard errors in the path
dispatch process (which would be costly), not to mention requiring
two path lookups in the general case.

SHOVE solves this by evaluating the right hand "enfixee" only once:

   1 + 2 -> lib/* 3

In the above example, when the 2 is reached it uses its lookahead step
to see the SHOVE operator.  SHOVE then does a single fetch process of
the right hand side, and determines the parameter convention it has
for its first argument.  It then uses that parameter convention to
gather more argument state.

SHOVE allows PATH!, WORD!, GROUP!, and literal ACTION! on the right
hand side.  It allows the action to be prefix and will just treat it
as enfix, sending it its first parameter from the left.

For this first implementation, the label of the function is lost
when dispatching through a path.  There are also general issues of
quoting contentions between leftward things that quote right and
rightward things that quote left.