Preliminary implementation of JS-NATIVE
This is a first cut at adding a "JavaScript extension" which provides
a way to make "JavaScript natives".  These have specs which are BLOCK!
but bodies which are a TEXT! of code that will run in JavaScript.

While the question of how to extract the arguments from such functions
is open, this adds a special `rebArg()` function for the purpose.