Fix unzip to use inflate/max vs. inflate/limit, adjust test
The name "/limit" for the refinement on DECOMPRESS was ambiguous w.r.t.
/PART, because it wasn't clear if it was a maximum decompression or
how much of the data to compress.  It was changed to /MAX but this
change did not make it to %unzip.r.  As a result, UNZIP was no longer
working.

This wasn't revealed by the small unzip test, because that test was not
triggering the DEFLATE algorithm on the small data it was using.  This
fixes the bug, and adjusts the test to use a larger file that can
actually be compressed.