Fix DEFLATE bug when trimming oversized compression buffers
When DEFLATE is used to compress data, it guesses at a buffer size
based on the "deflateBound()".  If the data compresses significantly
better than that (e.g. > 1024 bytes of spare space in the buffer), it
reallocates the buffer to avoid waste.

There was a bug in the reallocation, so instead of truncating to the
actual size of the data (strm.total_out) it was truncating to the
excess space in the buffer (strm.avail_out).  If this quantity was
smaller than the compressed size, then this would result in an error.