Simplified `/`-as-zero-length-path implementation
This changes `/` from being a WORD! to being a zero-length path.  It
also makes `foo/` the representation of a length-1 path, which acts
the same as whatever the single element would be if not in a path.

(Reasoning for a length-1 path acting like the element alone is to
simplify code which might put an ACTION! in the first slot of a PATH!,
and then progressively build a list of refinements...but wind up with
an empty list.  By allowing the path to run anyway, this means there's
no need to special-case that as just an ACTION!.)

The concept is that the zero length path dispatch in the evaluator goes
to the type on the left, just like in ordinary path dispatching.  But
it runs different code.  Hence `a/b` and `a / b` are two distinct ways
in which the value on the left is "pathed".  The left is known as
"path picking" while the right is tentatively called "path splitting".

In path picking, the right hand side is given quoted as-is to the
dispatcher (unless in a GROUP! or if it's a GET-WORD!).  But in path
splitting, the right hand side is given to the dispatcher after having
undergone "tight" evaluation.  When combined with the idea that a
number on the left will trigger division, this means that `/` can act
consistently with historical division in Rebol...despite not being
a WORD!.

As an added experiment, if the left hand side is an ANY-STRING! then
the code will run SPLIT with the right hand side.

   >> "abc;defghi;jkl" / ";"
   == ["abc" "defghi" "jkl"]

The implementation does not go formally through a type dispatcher the
way PD_Xxx() functions do for path dispatch presently.  Instead, there
is a native called PATH-0 which is invoked.  This provides the interim
hook point, which can be hijacked or overridden in usermode to add
support for new types if needed.