Allow library builds to have built-in extensions
The concept of a "built-in extension" was previously one in which the
extensions were considered to be built into the R3 executable.  This
meant that the required includes were in %host-main.c, and it would
deal directly at the C level...calling the internal API to convert
the raw C pointers into handles that could be loaded at the appropriate
moment by the startup code.

(One could imagine--for instance--a command line switch which asked
not to initialize a built-in extension.  Since the command line is
processed by Rebol, rebStartup() has to be run before that but can't
run the extension initializations.)

However, when building Rebol as a library it may be desirable to have
extensions ship with that library.  (There isn't much other meaning to
"built-in extension" for a library otherwise...it is the only product,
so there's nothing to build into.)  This situation arose with the wish
to offer JavaScript natives via an extension when building libr3.js.

To faciliate that, this changes the previous internal extension APIs
Prepare_Boot_Extension() and Shutdown_Boot_Extensions() to be
encapsulated in the libRebol API.  Rather than the caller having the
pointers to the C init functions and passing them in, it is assumed
that %a-lib.c itself already has awareness of the boot extensions and
has an API for offering them back as Rebol HANDLE! arrays.

Included in this is moving away from the idea that clients of libRebol
need to pass in any kind of "Host Lib" pointer.  Instead it is assumed
that the library is built and linked with enough knowledge of the
platform it runs on to not need that passed in.  Moving away from the
legacy Host_Lib has been an ongoing process, but this removes the
concern from the API level.