Patch bigint overflow, API handle leak in crypto module
When updating axTLS files, a patch for numeric overflow seems to be
omitted, which was causing problems in server key exchanges.  It isn't
clear at this moment if that means the "newer" axTLS was actually older
or if this is an isolated detail that a third party had noticed and
fixed.

Includes a fix for missing rebRelease() calls in the crypto module.