Fix HMAC-SHA256, memory leak in WAIT edge case, compiler warning
This fixes some issues found while doing the TLS 1.2 update.

The HMAC-SHA256 function had atrophied; e.g. was using LENGTH instead
of LENGTH OF.

If a WAIT wound up waiting on a BLOCK! of ports, with no ports needing
a wait, and no timeout, there was a memory leak of a temporary array
series produced by REDUCE'ing the argument to WAIT.

Some compiler settings were not accepting the `fallthrough;` trick,
so this switches to using explicit GOTO in switches in the scanner.