Lie about list of hash extensions in TLS ClientHello
In the TLS protocol, when a connection request is sent from a client
to a server it can give a list of "extensions" it supports.  The
server then replies about whether it supports those extensions or not.

It is technically possible for a server to refuse connection from a
client that does not support an extension it requires.  One such
situation is servers that check to see if the client has the needed
signature algorithms for checking its certificate.

Since the current TLS implementation does not check certificates,
there's not much point in not saying that the client can support
pretty much any signature algorithm.  This avoids the disconnection
from servers that are wired in such a way to not talk to clients that
can't check them.