Add Server Name Indication extension to TLS requests
If the host-name the TLS was asked to connect to was a string, then
this change causes the host name to be sent as a "SNI" extension:

https://en.wikipedia.org/wiki/Server_Name_Indication

This is important if the server is trying to host multiple secure
domains from the same IP address.  If so, the TLS protocol must know
which domain you are asking for, in order to know which certificate
to give you.

This is an extension to the TLS protocol, not something built into the
protocol itself.  However, many browsers and tools (like curl) have
implemented it.  Some websites will refuse the ClientHello request
if the extension is not present.