Adjust short-circuit AND/OR, temporarily disable GROUP!
The policy for using a BLOCK! on the right hand side of an AND/OR was
that they would act like ALL and ANY.  This is because instead of
"flattening" the output to just a LOGIC!, it was desirable to be able
to use the operation to produce results:

    x: some expression or [
       code that produces value for x
    ]

However, acting exactly like ALL or ANY meant returning NULL in the
case of failure, instead of falsey things.  Hence you would not be able
to write:

    some-flag: block? y and [length of y = 10]

The NULL result prohibited it.  Given this weakness of the form, the
idea was to use a GROUP! on the right hand side to force a LOGIC!.

This commit implements a different idea for short circuit AND and OR.
What it does is that if there is a falsey result which causes the
test to fail, then that falsey result is preserved verbatim.  And if
there is a truthy result that causes it to succeed, that result is
preserved verbatim:

    >> 10 and [null]
    == null

    >> 10 and [false]
    == #[false]

    >> 10 and [_]
    == _

This gives the appealing property that if you only put LOGIC! in, then
you will only get LOGIC! out:

    >> false and [true]
    == #[false]

    >> true and [false]
    == #[false]

A seeming sacrifice of this is AND/OR can't be used with NULL-activated
constructs, like ELSE or THEN.  However, that was a red herring, since
all falsey results were being flattened to NULL.  There is thus no
difference between `and [...] else [...]` vs. `and [...] or [...]`

This frees up the GROUP! right hand side semantics for potential use
as the bitwise/set-operator form.  One benefit of using the GROUP!
form for that is because group arguments imply evaluation before the
construct executes, which would conventionally prevent short-circuit.
That makes it seem a natural fit for something that must operate on
both arguments at once.  It would also avoid needing to come up with
a distinct name for the bitwise operators.