Get rid of fallthrough macro trick idea
An idea was tried which defined a macro called fallthrough to help make
it more explicit when fallthroughs happen.  However, various compilers
warn that a fallthrough might be implicit, and the most general way
to avoid that warning is to name the common point and use a `goto`.
This should not cost anything extra with optimization, and gives a
chance to make the reasoning for the commonality clearer by way of the
name given to the label.

Also moves toward use of the two-space indent for some switch labels,
as that increasingly seems to be a good compromise on the spacing.