Use C99 or shim `bool` instead of REBOOL
REBOOL has been a `#define` for bool, which is either the C99 stdbool.h
boolean or a shim version.  This goes ahead and does the search and
replace for that and TRUE and FALSE, so that the code is more in line
with standard C.

The TRUE and FALSE defines are used only in Windows-specific code,
where they come from Windows.h along with the BOOL type.