MAP-EACH/REMOVE-EACH return NULL on BREAK, no BREAK/WITH
This changes both MAP-EACH and REMOVE-EACH to fall in line with the
convention of returning NULL if a BREAK is encountered, and dropping
the in-progress partial operation:

    >> map-each x [1 2 3 4] [
           if x = 3 [break]
           x + 10
       ]
    ;-- null (would have historically been [11 12])

Now if an in-progress mapping wants to keep partial results, it would
need to do so manually by rejecting subsequent map items:

    >> skipping: false
    >> map-each x [1 2 3 4] [
           if skipping [continue]
           if x = 3 [skipping: true | continue]
           x + 10
       ]
    == [11 12]

*This is for a reason*:

The ability to be able to tell from outside a loop whether it had a
break or not is very important in building composite loop operations.
For instance, imagine a MAP-BOTH which took two MAP-EACH operations and
wanted to run them in sequence for two sets...but if the first map-each
were to encounter a break there'd be no straightforward way for the
MAP-BOTH operation to know that and avoid running the second MAP-EACH.

(It's also very useful, for casually signaling an ELSE or THEN outside
the loop...based on whether it finished or not.)

For these reasons (that have become increasingly important over time)
the /WITH refinement is now formally removed from BREAK.  CONTINUE
still has the refinement, which means to act as if the body end was
reached with the given value.

In the case of REMOVE-EACH, this currently means not doing the pending
removals if a break is encountered, because it would prevent returning
the number of elements that were removed.  This may not be the final
word on the design; it wouldn't seem as necessary if the return result
were just the original series.