CLOAK/DECLOAK->userspace %redbol.reb, kill CRC8/CRC16
The CLOAK and DECLOAK functions were used for a casual form of
XOR-based encryption, that was used to do some mild obfuscation of the
embedded code in Rebol.  There may be some historical Rebol scripts
dependent on the behavior, however they are not relevant forms of
modern encryption.

This implements a simple and unoptimized translation of CLOAK and
DECLOAK into userspace in the %redbol.reb compatibility layer, for
legacy usages...and removes them from the Crypt module.  Improvements
are left as an exercise to any stakeholder who wants to do anything
other than just be able to un-cloak some piece of data hanging around.

It also removes the CRC8 and CRC16 forms used by the Tello drone, as
userspace implementations for those have been written as well:

https://github.com/gchiu/Rebol3/blob/master/scripts/calcCRC8.reb
https://github.com/gchiu/Rebol3/blob/master/scripts/calcCRC16.reb