Remove extensions lists from VS configs
The extensions lists were changed to use MAKE MAP!.  The Visual Studio
configurations hadn't been updated for the change.

That change may not be an ideal one for a configuration file--and
should be rethought.  But for now, this removes the custom extensions
list (which was empty) so that the files can be used with the make
process.

Also gets rid of a stray ?! that hadn't been noticed in the ??/!!
deprecation.