Make rebSpell return a JavaScript 'string'
There was no code in the wrapper for variadics which would transform
the rebAlloc'd pointer from rebSpell() automatically into a Javascript
string.  This goes ahead and adds that, where it does the UTF8 to
JavaScript conversion and the rebFree() for you.

Also switches to not use semicolons, as that is the current operating
philosophy:

https://feross.org/never-use-semicolons/
https://standardjs.com/