Distinguish debug normal from debug symbols
Previously, the callgrind build was useful for when you wanted symbols
but no debugging information.  However, that also includes CALLGRIND
the controlling native for turning on and off statistics gathering.
It's typical to want symbols in an emscripten build without all the
asserts and debug features, so this makes `debug=symbols` mean just
that, while `debug=normal` is symbols and asserts.

Also includes a bugfix for when empty variadics are executed, found
during work on the JavaScript build.