Fix TIGHTEN for usermode functions in debug builds
Tightening makes a copy of a function's DETAILS (the internal array
which carries its implementation information).  Whether it should do
this or not is one question, but so long as it does make a copy it
shouldn't assume the cells are non-relative...which it was doing.

This removes the KNOWN() from TIGHTEN.

Includes trivial shortening of True/False initialization in boot.