Enable // for inline commenting
Rebol historically used `//` for a form of modulus.  Ren-C is turning
MOD into an enfix and tighted synonym of MODULO, and REM into the enfix
and tightened synonym for REMAINDER.  And with `/` being turned into
the notation for a zero-length PATH! (and delegating `a / b` to be a
special behavior dispatched similarly to how `a/b` is dispatched),
there is little good case to be made for having natural WORD!s with
slashes in them.

Hence while it may have been technically possible to say that
`value1 // value2` would perform a modulus operation, that is much
clearer as `value1 mod value2` or `value1 rem value2`...and consistent
with Haskell and Clojure (for example).

The widespread use of // for comments in many languages strongly
suggests that its success is not merely a y-product of chance.  The
C language is very hesitant to borrow from things found in C++ -- for
example -- but this form of comment was formally adopted in C99 (and
informally adopted much earlier).  It looks more "solid" and is not
broken up in a way that interferes with text the way the "messy" and
dotted pattern of `;;` does.

Additionally, when mixing fragments of code between JavaScript and
Rebol, being able to use one type of comment is helpful.  It will
almost certainly be preferred by many users of that mixture (including
the author of this commit).

This does not affect the existing usage of ; for comments, and it only
applies when // is at the start of a token sequence.  Hence a//b is
interpreted as a PATH! with a REFINEMENT! in the second slot (for
example).  However, `a///b` is an error.