Workaround for 0-length PATH! passed used with ME, etc.
Now that `/` is mechanically a 0-length path, the evaluator handles
the idea that is a synonym for division.  However, this doesn't make
things that previously depended on `get '/` returning a divide function
work.  For instance:

    x: 1
    x: me / 2

Mechanically, it may be that GET '/ needs to have an answer which
gives back the divide action.  This would not require GET to take a
new type (it already takes PATH!) but would present a bit of an
oddity since it could not be overridden.

For now, this does a small internal hack just to make the above case
work.