More elegant fix for 0-length PATH! problem than last commit
The previous commit was a targeted hack that made functions like ME
work with the 0-length path / (whatever it does).  This follow-up
is a more generalized fix, such that GET '/ will return the "PATH-0"
operation (what a 0-length PATH does).

Note that this operation need not be division, as demonstrated with
the idea that ("abcdef" / 2) is ["abc" "def"] ... so it is up to the
type on the left what it means, as with other forms of PATH! dispatch.