Workaround bug in old MinGW cross-compiled 64-bit longjmp
64-bit builds made by MinGW in the 4.x range have an unfortunate bug in
the setjmp/longjmp mechanic, which causes hangs for reasons that are
seemingly random, like "using -O0 optimizations instead of -O2".

Working around the problem involves using `__builtin_longjmp` and
`__builtin_setjmp` instead of the plain versions:

https://sourceforge.net/p/mingw-w64/bugs/406/

Bending to the bugs of broken compilers is usually not interesting, but
the Travis CI cross-platform builds on Linux targeting Windows were set
up on this old version--which otherwise is a good test the codebase
hasn't picked up dependencies that are too "modern".