const-correctness and type verification for TUPLE! macros
While the TUPLE! type is not really very solidly designed, it is still
around and in use.  The macros for accessing data in tuples were not
checking that the cell was a REB_TUPLE, and also did not have any
const-correctness.

This changes the C++ build to check these attributes.