Standardize IMAGE! underlying binary to RGBA byte order
R3-Alpha's implementation of the IMAGE! datatype used a series of
32-bit unsigned integers which contained pixel color values.  The
choice to address the data by unsigned integer instead of by individual
RGBA bytes made pixel addressing easier than using bytes in some cases,
as there was no need to involve multiplication by 4.  It also meant
that pixel colors could be easily passed by value instead of as an
array of 4 bytes.

There were several downsides to using integers, however.  One of which
is that the endianness of the platform would affect the byte ordering
of the pixels in the IMAGE! data format.  Another is that it prevented
the ability of the image data to participate in the rebMalloc() and
rebRepossess() strategy, since only byte-oriented buffers could be
retaken in this manner.

This commit goes ahead and canonizes the access to the pixels to use
access by byte, where images are assumed to always be in R G B A order
for their internal storage.  However, switching to use a byte-sized
series (instead of one with 4 bytes per element) would involve deeper
changes, since currently image runs through common series code that
implements operations like NEXT and BACK.  Since IMAGE! has a series
position (as currently written) that code must bump in 4 pixels per
unit, so the SER_WIDE() needs to be 4.

However, the changes in this commit eliminate a set of nuisance defines
that had been corraled into a file named "pixel-hack.h"