Shift Make_Image() routines to producing REBVAL, not REBSER
This changes the API for creating images to something which does not
expose that the full information for the image is linked to by a REBSER
node (e.g. having the width and height packed into the MISC() field).

It is a small step toward turning IMAGE! into something closer to a
user-defined type (instead of something more tightly baked into the
REBSER definition).