Don't count dummy stack frames in WHERE information
The call stack is produced by examining the list of frames on the stack
and adding an element to the block for each ACTION! invocation.

However, some frames are "dummy" frames which exist solely as a way
to aggregate API handle lifetimes for C code.  So libRebol API calls
(such as those in the console that come from main() to run lines of
code) are bracketed by what appears to be a function call even though
it is at the top level...so that handles allocated will be cleaned
up when an error is executed.

These dummy frames appeared as anonymous function calls in the stack
trace (currently shown as a BLANK!).  This commit filters those out,
and also makes it so that errors do not show any `Where:` information
when the stack trace is an empty block.

An effect of this is that an error typed in at the top level of the
console with no function call will not have a superfluous WHERE, and
avoids a situation where all error traces start with a BLANK!.

Also corrects a typo where `<void>` was changed to `<void!>`