Cleanup of /PART handling, fixes FIND/PART issue (others?)
For handling partial ranges, R3-Alpha had two helper routines called
Partial() and Partial1().  They were somewhat complex looking and had
significant overlap.  One of the trickiest aspects was that they could
mutate the incoming series argument to adjust its position to be the
beginning of a /PART range (if the range was specified by a negative
number, or by a position before the series location).

This commit merges the two routines and creates a few variants, whose
names indicate their semantics (e.g. Part_Len_May_Modify_Index).  This
makes it more obvious at the callsite when an index was used that had
been captured prior to the operation, and fixes some of those cases,
including one that caused a bug found recently.

It's likely that this change resolves some other bugs, but it certainly
makes what the code was doing clearer.