Sacrifice const_KNOWN() distinction to check only in C++ build
It's not possible in C to overload function signatures to take in
different datatypes, and return different datatypes.  This means that
a simple operation can't be defined which takes in either a const
and returns a const, or a non-const and returns a non-const.  That's
only possible with macros.

As a result, the KNOWN() operation came in a pair...const_KNOWN and
KNOWN.  However, if one is willing to only do const checking in the C++
build, then the C++ build can define an overloaded KNOWN() while the
C build can just discard const-correctness and always return a mutable
value, even if the input is const.

It's expected in general that the core must pass a C++ build, as there
are many other things checked (including the difference between RELVAL
and REBVAL, which are indistinct in the first place in C...and the
very reason for the existence of KNOWN (!)).  Hence, to simplify the
code at the callsites, this goes ahead and sacrifices the ability to
do const checking in the C build, collapsing const_KNOWN and KNOWN to
one operator that only does const checking in C++.