Kill of SINK() mechanic, simplify cell initialization
In the early creation of specific binding, the type distinction between
relative and specific values was mainted for values that were outputs
as well as inputs.  This meant that if something like Init_Integer()
were going to be initializing a cell as something that would not be
relative, its out parameter would be a REBVAL*.

However, if someone was passing in an uninitialized cell, they would
only have a RELVAL*...and the only way to promote it to a REBVAL* to
be used as an output was to have the bits run through a check that
ensured it was actually specific.  Since that wouldn't work, SINK()
was created as a way of promoting RELVAL* to REBVAL* without checking
them, and instead would guarantee the bits were uninitialized in the
debug build by writing them with garbage.

Over time, the inconvenience of having to SINK() things led routines
that did REBVAL* initializations to take RELVALs in, and give REBVAL*
out as the promoted pointer.  This was a lot more convenient.  Also,
SINK() was found to cause some amount of problems for the GC based on
how exactly it chose to "trash" the bits, so it had to become more
complex over time.

The number of usages of SINK() dwindled significantly, most immediately
followed by initializations that could provide the value itself.
This changes many of those instances to use a RESET_CELL() primitive,
which directly hands back a REBVAL* if the caller knows that what they
are assigning is specific.

By distinguishing a cell reset from a mere reset of a value's header,
tracking information is improved--even if the payload is overwritten
(due to the existence of DEBUG_TRACK_EXTEND_CELLS).  Also, by noticing
that RESET_CELL() returns a REBVAL*, initialization functions like
Init_Void() can be turned into simple macros--reducing the amount of
code in %sys-value.h noticeably.